+++ src/main/java/org/jenga/dantong/survey/repository/SurveyRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface SurveyRepository extends JpaRepository<Survey, Long> {
-    Survey findBySurveyId(int surveyId);
+    Survey findBySurveyId(Long surveyId);
 }
